Adding basic tests for failedSpecDetails
